Merge in content from `ruby-enumerable-builtins`
Combine the READMEs and notes for the two lessons, as well as the `lib`/`bin/`
files. I chose to merge built-ins into this lesson instead of the other way
around because the name still makes sense (this lesson uses two distinct
modules, Enumerable and Comparable), and because this repo had more open issues.